requirements.txt
blinker==1.9.0
certifi==2025.11.12
chardet==4.0.0
click==8.3.1
cloudinary==1.44.1
Flask==3.1.2
Flask-Admin==2.0.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyMySQL==1.1.2
python-dotenv==1.2.1
requests==2.25.1
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
WTForms==3.2.1
